FROMLIST: binder: protect transaction_stack with inner lock.
(from https://patchwork.kernel.org/patch/9817779/)

This makes future changes to priority inheritance
easier, since we want to be able to look at a thread's
transaction stack when selecting a thread to inherit
priority for.

It also allows us to take just a single lock in a
few paths, where we used to take two in succession.

Change-Id: Idb1b6e9faa5c669978b2b3011fe326be8aece586
Signed-off-by: Martijn Coenen <maco@google.com>
Signed-off-by: FireLord1 <firelord.xda@lineageos.org>
Signed-off-by: Subhaam <derksoul98@gmail.com>